- fix types exported from wrong file, causing HMR failure
